+++ changelog.md
@@ -35,3 +35,5 @@ Config the System version in **init**.py
 - Initial Development#7.1: background animation added on December 12, 2022 -> **0.421**
 
 - Initial Development#7.2: minor improvements on December 12, 2022 -> **0.422**
+
+- Initial Development#8: account security check added on December 16, 2022 -> **0.45**